switch to continous time

+++ pkg/core/edge.go
@@ -3,7 +3,7 @@ package core
 type Edge interface {
 	GetTarget() Node
 	GetSource() Node
-	GetDelay() int
+	GetDelay() float64
 	GetWeight() float64
 	UpdateWeight(deltaW float64) float64
 }
